+++ NEWS.md
@@ -1,3 +1,7 @@
+* changed function name `load_metadata()` is now `load_pkg_metadata()` so as not to conflict with `DPchecker::load_metadata()`
+* updated all function names to be snake_case
+* updated all file names to match function names
+* updated `get_data_packages()` to (when `force = FALSE`) check for newer versions of requested data packages and  prompt the user to download the newest version if they so choose.
 * updated `get_data_packages()` to:
   * include a force option. Force defaults to false for a verbose & interactive function. Setting force = TRUE removes the interactive portions and eliminates all informative messages except critical errors.
   * `get_data_packages()` now inspects the location the data packages are being written to. If a folder with the data package reference id already exists, the function will prompt the user asking if they want to re-download and overwrite the existing data package (when force = FALSE)